requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
psycopg2-binary==2.9.7
pandas==2.1.3
openpyxl==3.1.2
django-import-export==3.3.1
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
fcm-django==1.0.12
django-anymail==10.2
requests==2.31.0
Pillow==10.1.0
django-bootstrap5==23.3
django-leaflet==0.29.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
